+++ src/SOMcreator/filehandling/obj.py
@@ -8,8 +8,9 @@ from SOMcreator.filehandling import core
 
 if TYPE_CHECKING:
     from SOMcreator import Project
-    from SOMcreator.filehandling.typing import ObjectDict
+    from SOMcreator.filehandling.typing import ObjectDict,MainDict
 
+### Import ###
 
 def load_object(proj, object_dict: ObjectDict, identifier: str) -> None:
     name, description, optional, parent, filter_matrix = core.load_basics(proj,object_dict)
@@ -37,3 +38,33 @@ def load(proj: Project, main_dict: dict):
 
     for uuid_ident, entity_dict in objects_dict.items():
         load_object(proj, entity_dict, uuid_ident)
+
+### Export ###
+def create_object_entry(element: classes.Object) -> ObjectDict:
+
+    object_dict: ObjectDict = dict()
+    core.fill_basics(object_dict, element)
+
+    if isinstance(element.ifc_mapping, set):
+        object_dict[IFC_MAPPINGS] = list(element.ifc_mapping)
+    else:
+        object_dict[IFC_MAPPINGS] = list(element.ifc_mapping)
+
+    psets_dict = dict()
+    for pset in element.get_all_property_sets():
+        psets_dict[pset.uuid] = property_set.create_pset_entry(pset)
+
+    object_dict[PROPERTY_SETS] = psets_dict
+    object_dict[ABBREVIATION] = element.abbreviation
+
+    if isinstance(element.ident_attrib, classes.Attribute):
+        object_dict[IDENT_ATTRIBUTE] = element.ident_attrib.uuid
+    else:
+        object_dict[IDENT_ATTRIBUTE] = element.ident_attrib
+
+    return object_dict
+
+def save_objects(proj:Project,main_dict:MainDict):
+    main_dict[OBJECTS] = dict()
+    for obj in sorted(proj.get_all_objects(), key=lambda o: o.uuid):
+        main_dict[OBJECTS][obj.uuid] = create_object_entry(obj)
