+++ src/SOMcreator/filehandling/property_set.py
@@ -7,7 +7,7 @@ from SOMcreator.filehandling import attribute
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from SOMcreator.filehandling.typing import PropertySetDict
+    from SOMcreator.filehandling.typing import PropertySetDict, MainDict
     from SOMcreator import Project
 
 
@@ -21,9 +21,28 @@ def load_pset(proj: Project, pset_dict: PropertySetDict, identifier: str, obj: c
     SOMcreator.filehandling.parent_dict[pset] = parent
 
 
-def load_predefined(project: Project, main_dict: dict):
+def load_predefined(project: Project, main_dict: MainDict):
     predef_pset_dict = main_dict.get(PREDEFINED_PSETS)
     predef_pset_dict = dict() if core.check_dict(predef_pset_dict, PREDEFINED_PSETS) else predef_pset_dict
 
     for uuid_ident, entity_dict in predef_pset_dict.items():
         load_pset(project, entity_dict, uuid_ident, None)
+
+
+#### Export ####
+
+def create_pset_entry(pset: classes.PropertySet) -> PropertySetDict:
+    pset_dict: PropertySetDict = dict()
+    core.fill_basics(pset_dict, pset)
+    attributes_dict = dict()
+    for attrib in pset.get_all_attributes():
+        new_dict = attribute.create_attribute_entry(attrib)
+        attributes_dict[attrib.uuid] = new_dict
+    pset_dict[ATTRIBUTES] = attributes_dict
+    return pset_dict
+
+
+def save_predefined(proj: Project, main_dict: MainDict):
+    main_dict[PREDEFINED_PSETS] = dict()
+    for predefined_property_set in sorted(proj.get_predefined_psets(), key=lambda x: x.uuid):
+        main_dict[PREDEFINED_PSETS][predefined_property_set.uuid] = create_pset_entry(predefined_property_set)
